Merge to main in preparation for v1.1 release

+++ test/multi_lang/root_wrapper_test.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2022 by Camden Mannett.
+// Copyright (C) 2022-2023 by Camden Mannett.
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE or copy at https://www.boost.org/LICENSE_1_0.txt)
 
@@ -43,7 +43,7 @@ BOOST_AUTO_TEST_CASE(parse_test)
                 });
 
         try {
-            r.parse(args.size(), const_cast<char**>(args.data()));
+            r.parse(static_cast<int>(args.size()), const_cast<char**>(args.data()));
             BOOST_CHECK(exception_message.empty());
             BOOST_REQUIRE(!!result);
             BOOST_CHECK_EQUAL(*result, parse_result);
@@ -94,7 +94,7 @@ BOOST_AUTO_TEST_CASE(parse_default_test)
                 });
 
         auto args = std::vector{"foo", "--hello", "42"};
-        r.parse(args.size(), const_cast<char**>(args.data()));
+        r.parse(static_cast<int>(args.size()), const_cast<char**>(args.data()));
         BOOST_REQUIRE(!!result);
         BOOST_CHECK_EQUAL(*result, 42);
     }
